FRR: move the additional components to frr-tools
Because of the proliferation of the frr related tools that are starting
to pullute the root folder, we move all the frr addons under frr-tools.
